Fixed all delays and added sleep out booster on
I fixed all delays for the design and realized I missed a very small command at the beginning of the waveform.
That command was 'sleep out booster on'. I will do more research on the functionality of this command as well as all others soon.
The current example displays an all white frame to the 160 by 80 PMOD 0.96 SPI LCD screen.