Removed unnecessary BRAM block
I was using BRAM to keep track of how much parameters each command was using.
As an improvement I set the params_left register to the fixed number of parameters for the command type.
This removes a load of BRAM for future projects.
I also had to roll back to this version since the optimized demo I worked on seemed to not be working properly. I will revisit that at a later date.