Upload sending ACMD41 until OCR is READY
ACMD41 must be sent repeatedly until the OCR comes back from R3 response format with a 1 value in the 31st bit of the OCR.
I have done this with this example and can be seen by the PulseView screenshot provided.
I believe the next command I need to run is CMD2. I will do that sometime tomorrow.